Move some types into ast_types.rs
core and ast are too large to work in
